use validator

+++ src/main/java/com/anastasia/quotes_spring_boot_api/service/QuoteService.java
@@ -1,13 +1,15 @@
 package com.anastasia.quotes_spring_boot_api.service;
 
 import com.anastasia.quotes_spring_boot_api.dao.QuoteRepository;
-import com.anastasia.quotes_spring_boot_api.entity.Quote;
-import com.anastasia.quotes_spring_boot_api.rest.model.QuoteNotFoundException;
+import com.anastasia.quotes_spring_boot_api.model.dto.QuoteDTO;
+import com.anastasia.quotes_spring_boot_api.model.entity.Quote;
+import com.anastasia.quotes_spring_boot_api.model.exceptions.QuoteNotFoundException;
+import com.anastasia.quotes_spring_boot_api.utils.QuoteMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.Random;
 
 @Service
@@ -15,46 +17,60 @@ public class QuoteService {
 
     private final QuoteRepository quoteRepository;
 
+    private static final Random RANDOM = new Random();
+
     @Autowired
     public QuoteService(QuoteRepository quoteRepository) {
         this.quoteRepository = quoteRepository;
     }
 
-    public List<Quote> findAll() {
-        return quoteRepository.findAll();
+    public List<QuoteDTO> findAll() {
+        List<Quote> quotes = quoteRepository.findAll();
+        return QuoteMapper.INSTANCE.quotesToQuoteDTOs(quotes);
     }
 
-    public Quote findById(int id) {
-        Optional<Quote> result = quoteRepository.findById(id);
-        if (result.isPresent()) {
-            return result.get();
+    public QuoteDTO findById(int id) {
+        return quoteRepository.findById(id)
+                .map(QuoteMapper.INSTANCE::quoteToQuoteDTO)
+                .orElseThrow(() -> new QuoteNotFoundException("Did not find quote with id = " + id));
+    }
+
+    public QuoteDTO create(QuoteDTO quoteDTO) {
+        Quote quote = QuoteMapper.INSTANCE.quoteDTOToQuote(quoteDTO);
+        Quote savedQuote = quoteRepository.save(quote);
+        return QuoteMapper.INSTANCE.quoteToQuoteDTO(savedQuote);
+    }
+
+    @Transactional
+    public QuoteDTO update(QuoteDTO quoteDTO) {
+        Quote quote = QuoteMapper.INSTANCE.quoteDTOToQuote(quoteDTO);
+        if (!quoteRepository.existsById(quote.getId())) {
+            throw new QuoteNotFoundException("Quote id not found - " + quote.getId());
         }
-        else {
-            throw new QuoteNotFoundException("Did not find quote with id = " + id);
-        }
+        Quote savedQuote = quoteRepository.save(quote);
+        return QuoteMapper.INSTANCE.quoteToQuoteDTO(savedQuote);
     }
 
-    public boolean existsById(int id) {
-        return quoteRepository.existsById(id);
-    }
-
-    public Quote save(Quote quote) {
-        return quoteRepository.save(quote);
-    }
-
+    @Transactional
     public void deleteById(int id) {
+        if (!quoteRepository.existsById(id)) {
+            throw new QuoteNotFoundException("Quote id not found - " + id);
+        }
         quoteRepository.deleteById(id);
     }
 
-    public List<Quote> findQuotesByText(String text) {
-        return quoteRepository.findQuotesByText(text);
+    public List<QuoteDTO> findQuotesByText(String text) {
+        List<Quote> quotes = quoteRepository.findQuotesByText(text);
+        return QuoteMapper.INSTANCE.quotesToQuoteDTOs(quotes);
     }
 
-    public Quote findRandomQuote() {
-        List<Quote> quotes = findAll();
-        if (quotes==null || quotes.isEmpty()) {
+    public QuoteDTO findRandomQuote() {
+        List<Quote> quotes = quoteRepository.findAll();
+        if (quotes.isEmpty()) {
             throw new QuoteNotFoundException("Couldn't find any quote");
         }
-        return quotes.get(new Random().nextInt(quotes.size()));
+        Quote randomQuote = quotes.get(RANDOM.nextInt(quotes.size()));
+        return QuoteMapper.INSTANCE.quoteToQuoteDTO(randomQuote);
     }
+
 }
